Fix exception handling if delete fails
Fixes an issue where ospurge silently stopped deleting resources if
deleting one resource failed.

If there is an exception while deleting some resource and that exception
does not have an inner exception the exception handling raised another
exception and the error was not propagated to the ospurge return code.
No error appeared in the log either.